add new functionality to check before delete a board

+++ board/src/main/java/br/com/dio/board/ui/MainMenu.java
@@ -1,5 +1,7 @@
 package br.com.dio.board.ui;
 
+import br.com.dio.board.dto.BoardDetailsDTO;
+import br.com.dio.board.dto.BoardInfoDTO;
 import br.com.dio.board.persistence.entity.BoardColumnEntity;
 import br.com.dio.board.persistence.entity.BoardColumnTypeEnum;
 import br.com.dio.board.persistence.entity.BoardEntity;
@@ -124,9 +126,19 @@ public class MainMenu {
     private void deleteBoard() throws SQLException {
         System.out.println("Informe o id do board que será excluido");
         var id = scanner.nextLong();
+        scanner.nextLine();
 
         try(Connection connection = getConnection()){
             BoardService service = new BoardService(connection);
+            BoardInfoDTO dto = service.getInfo(id);
+
+            System.out.println("Digite o nome do board \"" + dto.name() + "\" para confirmar a exclusão");
+            var name = scanner.nextLine();
+            if(!name.equals(dto.name())) {
+                System.out.println("O nome digitado não está correto. Operação cancelada!");
+                return;
+            }
+
             if (service.delete(id)){
                 System.out.printf("O board %s foi excluido\n", id);
             } else {
